combined rec and nn task in a single file
